webapp/requirements.txt
Django==3.0.5
djangorestframework==3.11.0
django-markdown==0.8.4
django-filter==2.2.0
django-rest-swagger==2.2.0
djoser==2.0.3
Pillow==7.1.1
PyMySQL==0.9.3
mysqlclient==1.4.6
django-phonenumber-field==4.0.0
phonenumbers==8.12.1